Removed Youtube Media Hooking
This project has been merged into One Tap Video Download. Just for the
sake of the consistency as well as to continue the legacy, this *empty*
module will still be kept and users will have to install it to enable
youtube video downloading in the main project(One Tap Video Download).